+++ src/styles/artist-slide/Slide.module.css
@@ -1,3 +1,38 @@
+@-webkit-keyframes skeleton-gradient {
+  0% {
+      background-color: rgba(165, 165, 165, 0.1);
+  }
+
+  50% {
+      background-color: rgba(165, 165, 165, 0.3);
+  }
+
+  100% {
+      background-color: rgba(165, 165, 165, 0.1);
+  }
+}
+
+@keyframes skeleton-gradient {
+  0% {
+      background-color: rgba(165, 165, 165, 0.1);
+  }
+
+  50% {
+      background-color: rgba(165, 165, 165, 0.3);
+  }
+
+  100% {
+      background-color: rgba(165, 165, 165, 0.1);
+  }
+}
+.skeleton {
+  width: 244px !important;
+  height: 322px !important;
+  margin-bottom: 20px;
+  border-radius: 5px;
+  -webkit-animation: skeleton-gradient 1.8s infinite ease-in-out;
+  animation: skeleton-gradient 1.8s infinite ease-in-out;
+}
 .container {
   width: 375px;
   margin-left: -45px;
@@ -7,6 +42,7 @@
   width: 244px !important;
   height: 322px !important;
   margin-bottom: 20px;
+  border-radius: 5px;
 }
 
 .cardImg {
